+++ packer_provisioning_scripts/ubuntu_2004_2204_preinstall.sh
@@ -3,6 +3,9 @@
 echo "$(date): ### Starting ubuntu_2004_2204_preinstall.sh ###"
 
 # parsing command line options
+INSTALL_CODE_DEPLOY_AGENT=false
+INSTALL_EFS_UTILS=false
+USE_GRAVITON=false
 while [[ $# -gt 0 ]]; do
     key="$1"
 
@@ -77,7 +80,9 @@ fi
 
 # install efs mount helper - requires git
 if [ "$INSTALL_EFS_UTILS" = true ]; then
-    apt-get -y install binutils git
+    curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -q -y
+    source /root/.bashrc
+    apt-get -y install git binutils rustc cargo pkg-config libssl-dev
     git clone https://github.com/aws/efs-utils /tmp/efs-utils
     cd /tmp/efs-utils
     ./build-deb.sh