+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Adding cargo to fix efs-utils install
+
 # 1.4.0
 
 * Integrating graviton preinstall script with main preinstall script